+++ app/Http/Livewire/FavoriteProductShow.php
@@ -24,15 +24,23 @@ class FavoriteProductShow extends Component
 
     public function store()
     {
-        $fav = favoriteProduct::where('product_id', $this->product->id)->where('user_id', auth()->id())->first();
 
-        if ($fav === null) {
-            favoriteProduct::create([
-                'product_id' => $this->product->id,
-                'user_id' => auth()->id()
-            ]);
+
+        if (auth()->id() != null) {
+            $fav = favoriteProduct::where('product_id', $this->product->id)->where('user_id', auth()->id())->first();
+
+            if ($fav === null) {
+                favoriteProduct::create([
+                    'product_id' => $this->product->id,
+                    'user_id' => auth()->id()
+                ]);
+            } else {
+                $fav->delete();
+            }
         } else {
-            $fav->delete();
+            $this->redirect(route('login'));
         }
+
+
     }
 }
